fix: restore security defaults and add organization directory Add a dedicated Organizations directory so users with access to multiple organizations can choose where to work before opening organization-specific people, reports, rooms, or settings. Move organization listing/search/create and removal flows out of the People tab, add directory navigation from the dashboard and organization pages, and surface parent/sub-organization context on the organization overview. Restore several security and compatibility defaults that had regressed: keep daily_use cross-user access limited to admin_support_actions, use the correct Sprockets action_cable asset name, discover the latest gitleaks release dynamically in CI, and restore the prior registration role ordering and default English labels. Update the CSP compliance note to describe the active report-only configuration and report collection endpoint. Also update the daily_use controller spec so supervisors with supervise-only access remain unauthorized for detailed daily usage logs.
