+++ .github/workflows/ci.yml
@@ -158,8 +158,7 @@ jobs:
           fetch-depth: 0
 
       - name: Run gitleaks (CLI - no license required)
-        env:
-          GITLEAKS_VERSION: 8.21.2
         run: |
-          curl -sSfL "https://github.com/gitleaks/gitleaks/releases/download/v${GITLEAKS_VERSION}/gitleaks_${GITLEAKS_VERSION}_linux_x64.tar.gz" | tar -xz gitleaks
+          GITLEAKS_DOWNLOAD_URL="$(curl -sSfL https://api.github.com/repos/gitleaks/gitleaks/releases/latest | python3 -c "import json, sys; release=json.load(sys.stdin); print(next(asset['browser_download_url'] for asset in release['assets'] if asset['name'].endswith('_linux_x64.tar.gz')))")"
+          curl -sSfL "$GITLEAKS_DOWNLOAD_URL" | tar -xz gitleaks
           ./gitleaks detect --source . --log-opts="--all" --exit-code 1 || true
