+++ app/assets/javascripts/application.js
@@ -1,5 +1,5 @@
 //= require application-preload.js
-//= require actioncable
+//= require action_cable
 //= require vendor.js
 //= require frontend.js
