+++ app/frontend/app/router.js
@@ -91,6 +91,7 @@ Router.map(function() {
   this.route('start_codes', { path: '/start_codes/:id' });
   this.route('admin', { path: '/admin' });
   this.route('lesson', { path: '/lessons/:lesson_id/:lesson_code/:user_token'});
+  this.route('organizations', { path: '/organizations' });
   this.route('organization', { path: '/organizations/:id' }, function() {
     this.route('people');
     this.route('reports');
