+++ app/frontend/app/templates/organization.hbs
@@ -40,6 +40,9 @@
           </span>
         {{/if}}
       </div>
+      <div class="md-org-hero-actions">
+        <LinkTo @route="organizations" class="md-org-hero-actions__link">{{t "Switch organization" key="switch_organization"}}</LinkTo>
+      </div>
     </header>
 
     <main class="md-main md-main--org">
