+++ public/locales/en.json
@@ -15,13 +15,13 @@
   "phrase_based": "Phrase-Based",
   "keyboards": "Keyboards",
   "pick_type": "- Choose your Role -",
-  "registration_type_communicator": "Communicator",
-  "registration_type_parent_communicator": "Parent and communicator",
-  "registration_type_parent": "Supervising parent",
-  "registration_type_slp": "Therapist",
-  "registration_type_teacher": "Teacher",
-  "registration_type_eval": "Evaluation/assessment device",
-  "registration_type_other": "Aide, caregiver or other supporter",
+  "registration_type_communicator": "A communicator",
+  "registration_type_parent_communicator": "A parent and communicator",
+  "registration_type_parent": "A supervising parent",
+  "registration_type_slp": "A therapist",
+  "registration_type_teacher": "A teacher",
+  "registration_type_eval": "An evaluation/assessment device",
+  "registration_type_other": "An aide, caregiver or other supporter",
   "unspecified_empty": "[  ]",
   "level_1": "1 - Minimal Targets",
   "level_2": "2 - Basic Core",
